+++ app/models/ocom_products_cart.rb
@@ -1,0 +1,2 @@
+class OcomProductsCart < ApplicationRecord
+end