+++ db/migrate/20220714121437_change_fix_typo_ocom_stores.rb
@@ -1,0 +1,5 @@
+class ChangeFixTypoOcomStores < ActiveRecord::Migration[7.0]
+  def change
+    rename_column :ocom_stores, :defautl, :default
+  end
+end