+++ db/schema.rb
@@ -10,34 +10,37 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2022_07_13_131108) do
+ActiveRecord::Schema[7.0].define(version: 2022_07_14_121437) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "pgcrypto"
   enable_extension "plpgsql"
 
-  create_table "ocom_admin_users", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
-    t.string "email", null: false
-    t.string "password_digest", null: false
-    t.string "first_name", null: false
-    t.string "last_name", null: false
-    t.string "phone_number", null: false
-    t.datetime "confirmed_at"
+  create_table "ocom_products", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
+    t.string "name"
+    t.float "price"
+    t.string "image"
+    t.string "description"
+    t.string "category_id"
+    t.float "discount"
+    t.float "stock"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
   end
 
-  create_table "ocom_menus", force: :cascade do |t|
-    t.string "name"
-    t.bigint "parent_id"
-    t.string "redirect_to"
+  create_table "ocom_products_carts", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
+    t.string "cart_id"
+    t.string "product_id"
+    t.float "quantity"
+    t.float "price"
+    t.float "total_price"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
-    t.integer "store_id"
   end
 
   create_table "ocom_products_categories", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.string "name"
     t.string "store_id"
+    t.uuid "parent_id"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
   end
@@ -54,6 +57,7 @@ ActiveRecord::Schema[7.0].define(version: 2022_07_13_131108) do
     t.integer "content_menu_id"
     t.string "description"
     t.integer "owner_id"
+    t.boolean "default", default: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
   end