+++ db/seeds.rb
@@ -1,134 +1,45 @@
-OcomStore.create(
-    :name => "Ocom Default Store", 
-    :content_menu_id => 1, 
-    :description => "This is the first Ocom Store", 
-    :owner_id => 1,
-)
-#1
-OcomMenu.create(
-    :name => "Products",
-    :parent_id => 0,
-    :store_id => 1,
-)
-#2
-OcomMenu.create(
-    :name => "Clothes",
-    :parent_id => 1,
-    :store_id => 1,
-)
-#3
-OcomMenu.create(
-    :name => "Food",
-    :parent_id => 1,
-    :store_id => 1,
-)
-#4
-OcomMenu.create(
-    :name => "Electronics",
-    :parent_id => 1,
-    :store_id => 1,
-)
+OcomStore.create(:name=>"Open Commerce Default Store", :content_menu_id=>1, :description=>"Store 1", :owner_id=>1, :default=>true)
 
-#5
-OcomMenu.create(
-    :name => "About",
-    :parent_id => 0,
-    :store_id => 1,
-)
-#6
-OcomMenu.create(
-    :name => "Contact Us",
-    :parent_id => 0,
-    :store_id => 1,
-)
+OcomProductsCategory.create(:name=>"Shirts", :store_id=>1, :parent_id=>0)
+OcomProductsCategory.create(:name=>"T-Shirts", :store_id=>1, :parent_id=>0)
+OcomProductsCategory.create(:name=>"Pants", :store_id=>1, :parent_id=>0)
+OcomProductsCategory.create(:name=>"Shoes", :store_id=>1, :parent_id=>0)
+OcomProductsCategory.create(:name=>"Accessories", :store_id=>1, :parent_id=>0)
+OcomProductsCategory.create(:name=>"Jewelry", :store_id=>1, :parent_id=>0)
 
-OcomMenu.create(
-    :name => "Shirts",
-    :parent_id => 2,
-    :redirect_to => "/products/clothes/shirts",
-    :store_id => 1,
-)
-
-OcomMenu.create(
-    :name => "Pants",
-    :parent_id => 2,
-    :redirect_to => "/products/clothes/pants",
-    :store_id => 1,
-)
-
-OcomMenu.create(
-    :name => "Shoes",
-    :parent_id => 2,
-    :redirect_to => "/products/clothes/shoes",
-    :store_id => 1,
-)
-
-### FOOD
-
-OcomMenu.create(
-    :name => "Burgers",
-    :parent_id => 3,
-    :redirect_to => "/products/food/burgers",
-    :store_id => 1,
-)
-
-OcomMenu.create(
-    :name => "Pizzas",
-    :parent_id => 3,
-    :redirect_to => "/products/food/pizzas",
-    :store_id => 1,
-)
-
-### ELECTRONICS
-
-OcomMenu.create(
-    :name => "Laptops",
-    :parent_id => 4,
-    :redirect_to => "/products/electronics/laptops",
-    :store_id => 1,
-)
-
-OcomMenu.create(
-    :name => "Phones",
-    :parent_id => 4,
-    :redirect_to => "/products/electronics/phones",
-    :store_id => 1,
-)
+OcomProduct.create(:name=>"Product 1", :description=>"Product 1", :category_id=>1)
+OcomProduct.create(:name=>"Product 2", :description=>"Product 2", :category_id=>1)
+OcomProduct.create(:name=>"Product 3", :description=>"Product 3", :category_id=>1)
+OcomProduct.create(:name=>"Product 4", :description=>"Product 4", :category_id=>1)
+OcomProduct.create(:name=>"Product 5", :description=>"Product 5", :category_id=>1)
 
 
-
-OcomAdminUser.create(
-    :email => "example@gmail.com",
-    :password_digest => "pwd123",
-    :first_name => "John",
-    :last_name => "Doe",
-    :phone_number => "123-456-7890",
-)
-    
-
-OcomProduct.create(
-    :name => "Shirt",
-    :price => 10.00,
-    :description => "This is a shirt",
-    :image => "https://images.unsplash.com/photo-1518791841217-8f162f1e1131?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=800&q=60",
-    :store_id => 1,
-)
-
-OcomProduct.create(
-    :name => "Pants",
-    :price => 10.00,
-    :description => "This is a pants",
-    :image => "https://images.unsplash.com/photo-1518791841217-8f162f1e1131?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=800&q=60",
-    :store_id => 1,
-)
-
-OcomProduct.create(
-    :name => "Shoes",
-    :price => 10.00,
-    :description => "This is a shoes",
-    :image => "https://images.unsplash.com/photo-1518791841217-8f162f1e1131?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=800&q=60",
-    :store_id => 1,
-)
+OcomProduct.create(:name=>"Product 1", :description=>"Product 1", :category_id=>2)
+OcomProduct.create(:name=>"Product 2", :description=>"Product 2", :category_id=>2)
+OcomProduct.create(:name=>"Product 3", :description=>"Product 3", :category_id=>2)
+OcomProduct.create(:name=>"Product 4", :description=>"Product 4", :category_id=>2)
 
 
+OcomProduct.create(:name=>"Product 1", :description=>"Product 1", :category_id=>3)
+OcomProduct.create(:name=>"Product 2", :description=>"Product 2", :category_id=>3)
+OcomProduct.create(:name=>"Product 3", :description=>"Product 3", :category_id=>3)
+OcomProduct.create(:name=>"Product 4", :description=>"Product 4", :category_id=>3)
+
+
+OcomProduct.create(:name=>"Product 1", :description=>"Product 1", :category_id=>4)
+OcomProduct.create(:name=>"Product 2", :description=>"Product 2", :category_id=>4)
+OcomProduct.create(:name=>"Product 3", :description=>"Product 3", :category_id=>4)
+OcomProduct.create(:name=>"Product 4", :description=>"Product 4", :category_id=>4)
+
+
+OcomProduct.create(:name=>"Product 1", :description=>"Product 1", :category_id=>5)
+OcomProduct.create(:name=>"Product 2", :description=>"Product 2", :category_id=>5)
+OcomProduct.create(:name=>"Product 3", :description=>"Product 3", :category_id=>5)
+OcomProduct.create(:name=>"Product 4", :description=>"Product 4", :category_id=>5)
+
+
+OcomProduct.create(:name=>"Product 1", :description=>"Product 1", :category_id=>6)
+OcomProduct.create(:name=>"Product 2", :description=>"Product 2", :category_id=>6)
+OcomProduct.create(:name=>"Product 3", :description=>"Product 3", :category_id=>6)
+OcomProduct.create(:name=>"Product 4", :description=>"Product 4", :category_id=>6)